Manually Processed Raw data added
The folder "data_raw" is added. In this folder are csv files that are manually prepared from the UCI repository. Some pre-processing is done so that the datasets are easier to use. The links to each of the datasets are in "dataset-links.txt", and the modifications done are mentioned in the same file.